face_auth_system/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
face-recognition>=1.3.0
Pillow>=10.0.0
